Alters last migration to use bigint for gamecode

+++ even-better-server/db/migrate/20171030150019_add_gamecode_gametype_gamedate_to_bets.rb
@@ -2,12 +2,12 @@ class AddGamecodeGametypeGamedateToBets < ActiveRecord::Migration[5.1]
   def up
     add_column :bets, :game_date, :date
     add_column :bets, :game_type, :string
-    add_column :bets, :game_code, :integer
+    add_column :bets, :game_code, :bigint
   end
 
   def down
     remove_column :bets, :game_date, :date
     remove_column :bets, :game_type, :string
-    remove_column :bets, :game_code, :integer
+    remove_column :bets, :game_code, :bigint
   end
 end
